Create plots showing DV, PRED, and IPRED by individual

DV is plotted with a symbol while PRED and IPRED are plotted with lines and symbols. The plot is faceted by unique individual identifier (like ID or USUBJID) as well as other faceting variables.

dv_pred_ipred(
  data,
  id_per_plot = 9,
  facets = "ID",
  nrow = NULL,
  ncol = NULL,
  fun = NULL,
  ...,
  id_col = deprecated()
)

dv_pred_ipred_impl(
  data,
  x = pm_axis_time(),
  dv = pm_axis_dv(),
  pred = pm_axis_pred(),
  ipred = pm_axis_ipred(),
  facets = "ID",
  id_col = deprecated(),
  xbreaks = waiver(),
  xunit = pm_opts$time.unit,
  xlab = NULL,
  ylab = NULL,
  angle = NULL,
  strip.text = element_text(),
  font_size = deprecated(),
  margin = deprecated(),
  plot.margin = NULL,
  legend.position = "top",
  pred_lty = 2,
  ipred_lty = 1,
  pred_point = TRUE,
  ipred_point = TRUE,
  lwd = 0.5,
  size = pm_opts$scatter.size,
  dv_shape = 19,
  dv_line = FALSE,
  dv_lwd = 0.5,
  scales = "free",
  log_y = FALSE,
  use_theme = pm_theme(),
  dv_color = "black",
  ipred_color = "red2",
  pred_color = "blue2",
  ncol = NULL,
  nrow = NULL,
  axis.text.rel = NULL,
  fun = NULL
)

do_dv_pred_ipred(data, options = list())

Arguments

data

the data frame to plot.

id_per_plot

number of unique combinations of facets columns to include on each page.

facets

a character vector of column names to use for faceting the plot, passed to ggplot2::facet_wrap(); if passed in col_label() format, then the column data is modified with a call to glue::glue_data(), so that the strip label is more informative; see examples.

nrow

number of rows in the plot grid; passed to ggplot2::facet_wrap().

ncol

number of columns in the plot grid; passed to ggplot2::facet_wrap().

fun

a function accepting a gg object as argument and returning an updated gg object; if supplied, this function is applied to each plot in the output list.

...

additional arguments passed to dv_pred_ipred_impl().

id_col

deprecated; use facets argument instead.

x

the time-axis column, in col_label() format; the title portion is used for the x-axis title along with xunit; see also the xlab argument.

dv

the DV column, in col_label() format; the title portion is used for the y-axis title; see also the ylab argument.

pred

the name of the PRED column; col_label() format is allowed, but the label portion is discarded.

ipred

the name of the IPRED column; col_label() format is allowed, but the label portion is discarded.

xbreaks

x-axis (time) breaks; passed to ggplot2::scale_x_continuous().

xunit

used to form x-axis title only if xlab is not provided.

xlab

x-axis title; if not NULL, passed to ggplot2::xlab().

ylab

y-axis title; if not NULL, passed to ggplot2::ylab().

angle

rotation angle for x-axis tick labels; passed to rot_x().

strip.text

optionally, the result of ggplot2::element_text() to format the strip text (e.g. change the font size or padding).

font_size

deprecated.

margin

deprecated.

plot.margin

for the plot; passed ggplot2::theme().

legend.position

passed to ggplot2::theme().

pred_lty

PRED linetype; passed to ggplot2::geom_line().

ipred_lty

IPRED linetype; passed to ggplot2::geom_line().

pred_point

logical; should points be plotted for PRED?

ipred_point

logical; should points be plotted for IPRED?

lwd

line width for PRED and IPRED; passed to ggplot2::geom_line().

size

size of shape for DV, IPRED and PRED; passed to ggplot2::geom_point().

dv_shape

shape for DV; passed to ggplot2::geom_point().

dv_line

logical; if TRUE then a line is added to the plot connecting DV points.

dv_lwd

line width for DV; passed to ggplot2::geom_line() as lwd.

scales

passed to ggplot2::facet_wrap().

log_y

logical; if TRUE then y-axis is shown in log-scale.

use_theme

a theme to use for the plot.

dv_color

color to use for DV points.

ipred_color

color to use for IPRED line.

pred_color

color to use for PRED line.

axis.text.rel

relative text size for axis text; use this to selectively decrease font size for axis tick labels.

options

a named list of options to pass to dv_pred_ipred

Examples

data <- pmplots_data_obs()

p <- dv_pred_ipred(
  data,
  ylab = "Concentration (ng/mL)",
  nrow = 3, ncol = 3
)

p <- dv_pred_ipred(
  data,
  facets = c("ID", "STUDYc//Study: {STUDYc}")
)