Functions to generate axis data

Functions to generate axis data

pm_axis_tad(xunit = NULL)

pm_axis_time(xunit = NULL)

pm_axis_tafd(xunit = NULL)

pm_axis_res()

pm_axis_wres()

pm_axis_cwres()

pm_axis_cwresi()

pm_axis_npde()

pm_axis_pred(xname = NULL)

pm_axis_ipred(xname = NULL)

pm_axis_dv(yname = NULL)

Arguments

xunit

an optional time unit to be glued into the title.

xname

an optional name to be glued into the title.

yname

an optional name to be glued into the title.

Details

These functions call pmplots:::mk_col_title(). When time units (xunit) are passed, they will be wrapped in parens.

See also

Examples

pm_axis_time()
#> [1] "TIME//Time {xunit}"
pm_axis_time("h")
#> [1] "TIME//Time (h)"
pm_axis_pred("concentration")
#> [1] "PRED//Population predicted concentration"