Plots for pharmacometrics

Plots for pharmacometrics

Details

Help topics:

res_time Plots of residuals versus time

dv_pred Plots of observed data versus predictions

dv_time Plots of observed values versus time

res_pred Plots of residuals versus predictions

res_q Q-Q plots for residuals

eta_hist, eta_pairs, eta_cont, eta_cat Plots for ETA

split_plot split and plot

cont_cont Plots for continuous variables

layer Functions for adding layers to plots like reference or smoothing lines

Look and feel

See the pm_theme help topic for functions you can use to replicate the look and feel of plots generated by this package.