Plots for pharmacometrics

Plots for pharmacometrics

Look and feel

See the pm_theme help topic for functions you can use to replicate the look and feel of plots generated by this package.

Function listing

Residuals versus time

res_time, res_tad, res_tafd

wres_time, wres_tad, wres_tad

cwres_time, cwres_tad, cwres_tafd

Residuals versus PRED

res_pred, wres_pred, cwres_pred

Residuals versus covariates

res_cont, wres_cont, cwres_cont

res_cat, wres_cat, cwres_cat

NPDE plots

npde_time, npde_tad, npde_tafd

npde_pred

npde_hist

npde_q

npde_cont

npde_cat