Faceted plots

For these plots, data sets made long with respect to several y-axis variables and then plotted and faceted with ggplot2::facet_wrap.

wrap_cont_cont(
  df,
  x,
  y,
  ...,
  fun = pm_scatter,
  title = NULL,
  scales = "free_y",
  ncol = NULL,
  use_labels = FALSE,
  label_fun = label_parse_label
)

wrap_cont_time(df, ..., x = pm_axis_time())

wrap_res_time(df, ..., x = pm_axis_time())

wrap_eta_cont(df, x, y, scales = "fixed", ...)

wrap_hist(
  df,
  x,
  title = NULL,
  scales = "free_x",
  ncol = NULL,
  use_labels = FALSE,
  label_fun = label_parse_label,
  ...
)

wrap_dv_preds(
  df,
  ...,
  title = "Predicted {yname}",
  xname = "",
  scales = "fixed"
)

Arguments

df
data frame to plot

x
x-axis data in col_label format; if y has length greater than 1, then x must be length equal to 1

y
y-axis data in col_label format; if x has length greater than 1, then y must be length equal to 1

...
passed to fun

fun
the plotting function

title
a title to use for the axis with faceting groups

scales
passed to facet_wrap

ncol
passed to facet_wrap

use_labels
if TRUE, the label part of col_label will be used in the strip; the column name is used otherwise

label_fun
labeller function; passed to ggplot2::facet_wrap; the default is based on parse_label and allows latex markup in the label

xname
placeholder

Details

wrap_cont_cont is a general function used by the others to create faceted plots of two continuous variables. wrap_cont_time plots several continuous variables versus time. wrap_res_time plots several different residuals (or NPDE) versus time. wrap_eta_cont plots etas versus a continuous covariate. wrap_hist creates a faceted histogram plot.

For all plots, either x or y may contain multiple columns, but an error will be generated if both x and y list multiple columns.