Faceted plots

For these plots, data sets made long with respect to several y-axis variables and then plotted and faceted with ggplot2::facet_wrap().

wrap_cont_cont(
  df,
  x,
  y,
  ...,
  fun = pm_scatter,
  title = NULL,
  scales = "free_y",
  ncol = NULL,
  use_labels = FALSE,
  label_fun = deprecated(),
  labeller = label_tex
)

wrap_cont_time(df, ..., x = pm_axis_time())

wrap_res_time(df, ..., x = pm_axis_time())

wrap_eta_cont(df, x, y, scales = "fixed", ...)

wrap_hist(
  df,
  x,
  title = NULL,
  scales = "free_x",
  ncol = NULL,
  use_labels = FALSE,
  labeller = label_tex,
  label_fun = deprecated(),
  ...
)

wrap_dv_preds(
  df,
  ...,
  title = "Predicted {yname}",
  xname = "",
  scales = "fixed"
)

wrap_cont_cat(
  df,
  x,
  y,
  ...,
  title = NULL,
  scales = "free_y",
  ncol = NULL,
  use_labels = FALSE,
  labeller = label_tex,
  label_fun = deprecated()
)

Arguments

df

Data frame to plot.

x

x-axis data in col_label() format; if y has length greater than 1, then x must be length equal to 1.

y

y-axis data in col_label() format; if x has length greater than 1, then y must be length equal to 1.

...

Passed to fun.

fun

The plotting function.

title

A title to use for the axis with faceting groups.

scales

Passed to facet_wrap.

ncol

Passed to facet_wrap.

use_labels

If TRUE, the label part of col_label will be used in the strip; the column name is used otherwise.

label_fun

Deprecated; use labeller instead.

labeller

A labeller function; passed to ggplot2::facet_wrap(); the default is based on parse_label() and allows latex markup in the label.

xname

Placeholder.

Details

wrap_cont_cont is a general function used by the others to create faceted plots of two continuous variables. wrap_cont_time plots several continuous variables versus time. wrap_res_time plots several different residuals (or NPDE) versus time. wrap_eta_cont plots etas versus a continuous covariate. wrap_hist creates a faceted histogram plot. wrap_cont_cat plots continuous versus categorical data as a boxplot.

The following functions are called (as fun) to make each wrapped plot

wrap_cont_cont calls pm_scatter()

wrap_cont_time calls y_time()

wrap_res_time calls res_time()

wrap_eta_cont calls eta_cont()

wrap_hist calls cont_hist()

wrap_cont_cat calls pm_box();

For all plots, both x and y should name numeric data columns with the exception of wrap_cont_cat which expects x to name a categorical data column (the data are sent to pm_box()).

When pm_box() is called by wrap_cont_cat, the shown argument will be forced to the value FALSE.

For all plots, either x or y may contain multiple columns, but an error will be generated if both x and y list multiple columns.